update gitshrink.sh and install_dots.sh

+++ script/gitshrink.sh
@@ -21,27 +21,25 @@ done
 # @see http://stubbisms.wordpress.com/2009/07/10/git-script-to-show-largest-pack-objects-and-trim-your-waist-line/
 # @author Antony Stubbs
 
-# set the internal field spereator to line break, so that we can iterate easily over the verify-pack output
+# set the internal field separator to line break, so that we can iterate easily over the verify-pack output
 IFS=$'\n';
 
 # list all objects including their size, sort by size, take top 10
-objects=`git verify-pack -v .git/objects/pack/pack-*.idx | grep -v chain | sort -k3nr | head`
+objects=$(git verify-pack -v .git/objects/pack/pack-*.idx | grep -v chain | sort -k3nr | head)
 
 echo "All sizes are in kB. The pack column is the size of the object, compressed, inside the pack file."
 
 output="size,pack,SHA,location"
-for y in $objects
-do
-	# extract the size in bytes
-	size=$((`echo $y | cut -f 5 -d ' '`/1024))
-	# extract the compressed size in bytes
-	compressedSize=$((`echo $y | cut -f 6 -d ' '`/1024))
-	# extract the SHA
-	sha=`echo $y | cut -f 1 -d ' '`
-	# find the objects location in the repository tree
-	other=`git rev-list --all --objects | grep $sha`
-	#lineBreak=`echo -e "\n"`
-	output="${output}\n${size},${compressedSize},${other}"
+for y in $objects; do
+    # extract the size in bytes
+    size=$(( $(echo $y | cut -f 5 -d ' ') / 1024 ))
+    # extract the compressed size in bytes
+    compressedSize=$(( $(echo $y | cut -f 6 -d ' ') / 1024 ))
+    # extract the SHA
+    sha=$(echo $y | cut -f 1 -d ' ')
+    # find the objects location in the repository tree
+    other=$(git rev-list --all --objects | grep $sha)
+    output="${output}\n${size},${compressedSize},${other}"
 done
 
 echo -e $output | column -t -s ', '
@@ -49,8 +47,15 @@ echo -e $output | column -t -s ', '
 
 ## 3. Cleanup unnecessary files and optimize the local repository.
 
-git filter-branch --tag-name-filter cat --index-filter 'git rm -r --cached --ignore-unmatch filename' --prune-empty -f -- --all
-# After this command has finished executing, your repository should now be cleaned, with all branches and tags in tact.
+if command -v git-filter-repo &> /dev/null; then
+    echo "Cleaning the repository with git filter-repo..."
+    git filter-repo --path filename --invert-paths
+else
+    echo "You don't have git-filter-repo installed. Cleaning the repository with git filter-branch..."
+    git filter-branch --tag-name-filter cat --index-filter 'git rm -r --cached --ignore-unmatch filename' --prune-empty -f -- --all
+fi
+
+# After this command has finished executing, your repository should now be cleaned, with all branches and tags intact.
 
 
 ## 4. Reclaim space
